pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "enhanced-typing-assistant"
version = "1.0.0"
authors = [
    { name="Your Name", email="your.email@example.com" },
]
description = "An AI-powered typing assistant with accessibility features"
readme = "README.md"
requires-python = ">=3.8"
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: End Users/Desktop",
    "Topic :: Text Processing :: Linguistic",
    "Topic :: Office/Business",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Environment :: X11 Applications :: Qt",
    "Natural Language :: English",
]
dependencies = [
    "PyQt5>=5.15.9",
    "PyQt5-sip>=12.12.1",
    "openai>=1.12.0",
    "nltk>=3.8.1",
    "pyenchant>=3.2.2",
    "python-Levenshtein>=0.21.1",
    "pyttsx3>=2.90",
    "SpeechRecognition>=3.10.0",
    "sounddevice>=0.4.6",
    "soundfile>=0.12.1",
    "python-dotenv>=1.0.0",
    "cryptography>=41.0.0",
    "requests>=2.31.0",
]

[project.optional-dependencies]
test = [
    "pytest>=7.4.0",
    "pytest-qt>=4.2.0",
    "pytest-cov>=4.1.0",
]
dev = [
    "black>=23.3.0",
    "flake8>=6.0.0",
    "mypy>=1.3.0",
    "isort>=5.12.0",
]
docs = [
    "sphinx>=7.0.1",
    "sphinx-rtd-theme>=1.2.0",
]

[project.urls]
"Homepage" = "https://github.com/yourusername/typing_assistant"
"Bug Tracker" = "https://github.com/yourusername/typing_assistant/issues"

[tool.setuptools]
packages = ["typing_assistant"]

[tool.black]
line-length = 100
target-version = ['py38']
include = '\.pyi?$'

[tool.isort]
profile = "black"
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
line_length = 100

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true

[[tool.mypy.overrides]]
module = ["PyQt5.*"]
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "--verbose --cov=typing_assistant --cov-report=html"
